# Added logic to simplifier that creates equality out of > and <