@holycrap872 2015-04-20
Note, I update the commit on 4/20, adding more informative comments and making the logic a little easier to understand.