@holycrap872 2015-04-21
After a fair amount of testing, I found situations that cause assertion violations on what would become lines 230 and 240 in this patch.  They occur when signed comparator operations are being used.  The cause of the difficulty is that Klee represents all numbers as unsigned.  Therefore, when we have the operation (slt x 0), this is equivalent to (x <= -1).  Because Klee works in unsigned numbers however, it registers and overflow (the check 
'''
oneTooBig->compareContents(*constantValue) < 0
'''
on line 229 improperly returns true.

What I would like to do is to have a smarter check that, based on whether the comparator is signed or unsigned, chooses the correct compareContents function.  One of the compareContents would be a signed comparator, which the other would be unsigned.  Therefore, we could correctly identify overflow problems, without improperly triggering the assertion violation. I guess my question is does such a comparator exist, or would I need to create my own?

The following test case triggers the assertion on line 230.
'''

'''

Thanks.